net: ipv4: fragmentation: Drop the packet if fragmentation fails
If we could not fragment the IPv4 packet, then drop it and do
not try to send it. Let the upper layer re-send the packet if
needed. It is causing more trouble if we try to send the packet
and not honor the MTU setting.

Fixes #81021

(cherry picked from commit 414316379ad6604b39aa5a16632952ca0dd350d0)

Original-Signed-off-by: Jukka Rissanen <jukka.rissanen@nordicsemi.no>
GitOrigin-RevId: 414316379ad6604b39aa5a16632952ca0dd350d0
Cr-Build-Id: 8731572062094510337
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8731572062094510337
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Ic7707c3dfde4aa3f8782700292196b808acb3a47
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6008676
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>
Commit-Queue: Al Semjonovs <asemjonovs@google.com>