net: ipv6: fragmentation: Drop the packet if fragmentation fails
If we could not fragment the IPv6 packet, then drop it and do
not try to send it. Let the upper layer re-send the packet if
needed. It is causing more trouble if we try to send the packet
and not honor the MTU setting.

(cherry picked from commit fe56ce5a1c9659a05777754a2760866aad8e1fa3)

Original-Signed-off-by: Jukka Rissanen <jukka.rissanen@nordicsemi.no>
GitOrigin-RevId: fe56ce5a1c9659a05777754a2760866aad8e1fa3
Cr-Build-Id: 8731572062094510337
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8731572062094510337
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I27ada52dd62239a42f342291d971516beb966917
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6008677
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Commit-Queue: Al Semjonovs <asemjonovs@google.com>