net: utils: Port parsing failure in net_ipaddr_parse()
If trying to parse a string like 192.0.2.2:80/foobar and
setting the length to 12 which means to parse the IP address
and port, the parsing failed because it used one extra character
from the string. This issue was not present if the input string
was terminated after the port number.

Add a test case to catch this problem.

(cherry picked from commit 5f3e6212af7fefb86a081f148b3ef3991ba71a44)

Original-Signed-off-by: Jukka Rissanen <jukka.rissanen@nordicsemi.no>
GitOrigin-RevId: 5f3e6212af7fefb86a081f148b3ef3991ba71a44
Cr-Build-Id: 8730937883239782945
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8730937883239782945
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Iadf9429a68c3c7015fcf033c5fcd9d3ad94651d5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6028653
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: Tristan Honscheid <honscheid@google.com>
Tested-by: Tristan Honscheid <honscheid@google.com>
Reviewed-by: Tristan Honscheid <honscheid@google.com>