doc: twister: Clarify that the minimums are rough
There is no way to know for sure the size of an image built for a test,
given that zephyr can build for so many different architectures,
platforms, software supports, etc. So it should be clear that the filter
for min_ram and min_flash on the twister testcases is an estimate, and
not a strictly well known value.

(cherry picked from commit 39f32eac370249cfce99c8a0d150a8247b0b3a25)

Original-Signed-off-by: Declan Snyder <declan.snyder@nxp.com>
GitOrigin-RevId: 39f32eac370249cfce99c8a0d150a8247b0b3a25
Cr-Build-Id: 8731843854160875777
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8731843854160875777
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Ia126252dea2720cb9838914baec4c483a08d9c53
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6004806
Tested-by: Yuval Peress <peress@google.com>
Reviewed-by: Yuval Peress <peress@google.com>
Commit-Queue: Yuval Peress <peress@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>