modules: hal: nxp: update to include fix for CACHE64_GetInstanceByAddr()
CACHE64_GetInstanceByAddr() function was asserting when an instance was
requested for an invalid address that the CACHE64 controller does not
manage. This behavior is not correct, as the CACHE64 management
functions check to see if the instance number returned by this function
is out of range (and if so, simply return without modifying the cache).

This assertion was causing a failure within the USDHC driver, which
performs a cache clean/invalidate for tx/rx transfers within the HAL
layer. When a transfer was run using a data buffer not in the CACHE64
address range, this assertion failed and caused the application to crash

Fixes #80901

(cherry picked from commit 2ad01d966b7f5c499e3d52ffa37d75b40406bec7)

Original-Signed-off-by: Daniel DeGrasse <daniel.degrasse@nxp.com>
GitOrigin-RevId: 2ad01d966b7f5c499e3d52ffa37d75b40406bec7
Cr-Build-Id: 8731572062094510337
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8731572062094510337
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Ibcee7d60d3ab6faba3ff1555bbdc8fdd22c76146
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6008650
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>
Commit-Queue: Al Semjonovs <asemjonovs@google.com>