net: lib: coap_client: Don't clear internal structures on response
When response is received and handled, don't just clear the structure
but instead mark it as ongoing=false.
So if we later on receive a duplicate response for it, we can still
respond with Ack or Rst.

This is achieved by using release_internal_request() when we don't
expect any response for it and reset_internal_request() when we really
fill up a new request.

(cherry picked from commit 41ee35ae8b52095cce830ac3381b5ffa3c4162b4)

Original-Signed-off-by: Seppo Takalo <seppo.takalo@nordicsemi.no>
GitOrigin-RevId: 41ee35ae8b52095cce830ac3381b5ffa3c4162b4
Cr-Build-Id: 8731028481084229505
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8731028481084229505
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I3c16d6eb8e11d972cc0fa4b558892b594662cbb0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6028605
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: Tristan Honscheid <honscheid@google.com>
Tested-by: Tristan Honscheid <honscheid@google.com>
Reviewed-by: Tristan Honscheid <honscheid@google.com>