samples: Bluetooth: ISO: Update conn and adv intervals
Update the connection and advertising intervals of the
broadcast and connected ISO samples (including benchmark
samples) to work better with the selected SDU intervals
and the resulting ISO intervals.

For the ISO connected benchmark the order of CIG and ACl
has changed, so that we create the CIG before connecting
the ACL for the purpose of providing as much information
as possible to the controller.

(cherry picked from commit 420cafc1f54a8498b505abf38094467685857efc)

Original-Signed-off-by: Emil Gydesen <emil.gydesen@nordicsemi.no>
GitOrigin-RevId: 420cafc1f54a8498b505abf38094467685857efc
Cr-Build-Id: 8730286630880259329
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8730286630880259329
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I3c11ea5402a98eb530392b7672e3bedce5553db3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6048268
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>