samples: bt: reduce system heap memory for the hci_ipc sample
Moving from TinyCrypt to PSA Crypto API caused failures in the cpu_net
build due to RAM being overflowed. It turned out that 8192 bytes were
allocated for system heap memory, but Mbed TLS is the only user
of that memory (I found this though puncover) for AES purposes.
We reduce that to 1024 bytes because this should be enough for
this purpose.

in other samples/tests and babblesim, so a failure in building it
propagates in a lot of other failures.

(cherry picked from commit 184c0f9e9df4fa26e0508edff5bbb19e4790e850)

Original-Note: albeit this is also a standalone example, it's used extensively
Original-Signed-off-by: Valerio Setti <vsetti@baylibre.com>
GitOrigin-RevId: 184c0f9e9df4fa26e0508edff5bbb19e4790e850
Cr-Build-Id: 8730031914779849585
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8730031914779849585
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I4b60be4cb875401ad77e9831d658717c7bc76e8a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6057852
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>