net: prometheus: counter: Add function to set the counter value
This is similar function as add, but will add the difference of
previous value and the new one. This can be used if we want to
periodically update the value with a new one but don't want to
keep track of the old one.

(cherry picked from commit 0ec8385bec57432b364a1c67605c8f90e539d20b)

Original-Signed-off-by: Jukka Rissanen <jukka.rissanen@nordicsemi.no>
GitOrigin-RevId: 0ec8385bec57432b364a1c67605c8f90e539d20b
Cr-Build-Id: 8730164773885449249
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8730164773885449249
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I8b7dc2a1d2352c1954df881fd3d4c542769bf7e4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6053081
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>