tests: Bluetooth: Audio: Use same recv_cb for all tests
This commit changes the BSIM tests to use the same recv callback
for all tests. The purpose of this is to reduce code duplication
and make it easier to maintain the tests.

This also changes the recv_cb so that in case of any error we log
the most recently received SDU, which should provide more
information about why a test failed in case of RX error.

PBP had to be updated a bit to support the audio_stream
struct.

Also modifies a check and log in bap_stream that was less than
helpful to determine if it was the stream or the endpoint that
was NULL.

(cherry picked from commit b4f3763c3105e07a0e24a956e5725ba4d3c0c7cf)

Original-Signed-off-by: Emil Gydesen <emil.gydesen@nordicsemi.no>
GitOrigin-RevId: b4f3763c3105e07a0e24a956e5725ba4d3c0c7cf
Cr-Build-Id: 8730286630880259329
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8730286630880259329
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I0a767a75bc9eca03982491dbb8a343bf969586c9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6043646
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>