tests/bsim: Also runtime test the UART drivers for the nrf54l15
To increase coverage.

Note that we call twister separatedly for this target, due to the 54l15
overlays using UARTE20 which is indexed in simulation as "2", but as
the nrf52833 does not have 3 instances. So "-uart2_loopback" is not
a valid option for a 52833 executable.

(cherry picked from commit 4558056f81fe7387e6c291e3321dc880f1858ee2)

Original-Signed-off-by: Alberto Escolar Piedras <alberto.escolar.piedras@nordicsemi.no>
GitOrigin-RevId: 4558056f81fe7387e6c291e3321dc880f1858ee2
Cr-Build-Id: 8730286630880259329
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8730286630880259329
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Ic239a6344f401dc489963794b61c9dd0cb01016b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6043674
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>