kernel: Decouple sleep from suspend
Sleeping and suspended are now orthogonal states. That is, a thread
may be both sleeping and suspended and the two do not interact. One
repercussion of this is that suspending a thread will no longer
abort its timeout.

Threads are now created in the 'sleeping' state instead of a
'suspended' state. This dovetails nicely with the start delay that
can be given to a newly created thread--it is as though the very
first operation that a thread with a start delay is a sleep.

(cherry picked from commit 35435928c2f44b5bb2027da1e33c770108e8b97d)

Original-Signed-off-by: Peter Mitsis <peter.mitsis@intel.com>
GitOrigin-RevId: 35435928c2f44b5bb2027da1e33c770108e8b97d
Cr-Build-Id: 8728129377422644657
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8728129377422644657
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I3016475af09932b6d64a6aff84eb22765404ce8b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6110958
Reviewed-by: Dawid Niedźwiecki <dawidn@google.com>
Tested-by: Dawid Niedźwiecki <dawidn@google.com>
Commit-Queue: Dawid Niedźwiecki <dawidn@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>