net: http_server: fix header capture on concurrent http2 streams
Concurrent HTTP POST requests on different HTTP2 concurrent streams
require that the client's header_capture_context is re-used to capture
headers on a second stream before all of the body data has been received
(and sent to the application) on the first stream.

As a result, any captured headers must be sent to the application
callback before any headers can be received on a different stream. In
practice this means that for HTTP2 the application callback is called
for the first time on receiving a headers frame, before any data frames
are received. All subsequent application callbacks will not include the
request header data.

While this mechanism is not necessary for HTTP1, it is also updated to
only send headers in the first application callback for consistency.

Fixes #82273

(cherry picked from commit ddaeb1379a00bd4695a74e9f6eef0e189f8cd059)

Original-Signed-off-by: Matt Rodgers <mrodgers@witekio.com>
GitOrigin-RevId: ddaeb1379a00bd4695a74e9f6eef0e189f8cd059
Cr-Build-Id: 8729397734819344369
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8729397734819344369
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I6cb4560eaed3f1e5063f3d3d777cdd15ed76cd33
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6074809
Reviewed-by: Keith Short <keithshort@chromium.org>
Tested-by: Keith Short <keithshort@chromium.org>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Commit-Queue: Keith Short <keithshort@chromium.org>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>