twister: testplan: do not error on missing or unsatisfied env
bsim requires a special setup for it to be usuable and run by twister.
If this environment is not available, we should not error when the
platform is an integration platform.

Most developers and users of twister will not have the babblesim setup
and reporting errors is misleading and confusing. So if such env. is not
available, just filter as usual and do not error.

(cherry picked from commit e873d503a908ab5a313032bf59c7edfe9096a002)

Original-Signed-off-by: Anas Nashif <anas.nashif@intel.com>
GitOrigin-RevId: e873d503a908ab5a313032bf59c7edfe9096a002
Cr-Build-Id: 8729397734819344369
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8729397734819344369
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I110d4ec3cabcfebec79596a79448a2c002c74c25
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6074832
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: Keith Short <keithshort@chromium.org>
Reviewed-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Keith Short <keithshort@chromium.org>