usb: device_next: Update IAD first interface on init
IAD must be before the interfaces it associates and therefore there is
no need for the class to be in charge of updating the bFirstInterface.
Update IAD in common initialization code and remove the updates from
classes.

This fixes UAC2 instances where the IAD bFirstInterface is not 0, e.g.
when HID was used together with UAC2.

(cherry picked from commit 9006780db472470b4be12646e58446789341b427)

Original-Signed-off-by: Tomasz Moń <tomasz.mon@nordicsemi.no>
GitOrigin-RevId: 9006780db472470b4be12646e58446789341b427
Cr-Build-Id: 8728672958632284305
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8728672958632284305
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Ic91757b706e2aa5b52c60b8d038b84e8488c34c8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6093771
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Tested-by: Jeremy Bettis <jbettis@chromium.org>