usb: device_next: rework wSequence check in CDC NCM implementation
The consequences of a wSequence mismatch are unspecified and are
primarily for debugging purposes. Our checks are a bit too strict, and
if the header check fails, the sequence is not updated, and the header
check subsequently fails. Rework the code to just log a warning on
mismatch and also reset OUT sequence counter the same way as the IN
sequence counter.

(cherry picked from commit dd5492b3ce0fee61fa2382374f4c34e15ded7b76)

Original-Signed-off-by: Johann Fischer <johann.fischer@nordicsemi.no>
GitOrigin-RevId: dd5492b3ce0fee61fa2382374f4c34e15ded7b76
Cr-Build-Id: 8728129377422644657
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8728129377422644657
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I76e027f5eb1bf9e075e057922405cee5730177db
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6110908
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: Dawid Niedźwiecki <dawidn@google.com>
Reviewed-by: Dawid Niedźwiecki <dawidn@google.com>
Tested-by: Dawid Niedźwiecki <dawidn@google.com>