scripts: coredump: Fix loading coredumps when thread info is enabled.
When loading a coredump with an an elf built using
'CONFIG_DEBUG_THREAD_INFO=y', gdbstubs assumes that the thread info
memoryblock populated by 'CONFIG_DEBUG_COREDUMP_MEMORY_DUMP_THREADS=y'
will be present in the coredump. This is not always true and causes an
error, and a failure to load the coredump. Add a default value for the
threads_metadata variable in CoredumpLogFile which can be used to detect
when the memory block is not present. This allows the coredump to load
successfully.

(cherry picked from commit 9cdb07cb5291e6233dc32b8eecd0163944ae3264)

Original-Signed-off-by: Félix Turgeon <felixturgeon@meta.com>
GitOrigin-RevId: 9cdb07cb5291e6233dc32b8eecd0163944ae3264
Cr-Build-Id: 8728944750910134129
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8728944750910134129
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Ie5f88f718f4a5695a7945da45b852c3b2da77447
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6084325
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Tested-by: Jeremy Bettis <jbettis@chromium.org>