soc/mediatek: Add second "DMA" memory region
The Linux SOF driver owns this space, but puts buffers here which the
DSP is expected to read and write.  Give it a space in the default MPU
mapping.

Note that this assumes that the DMA region is numerically above the
linkable DRAM region in the address space.  That's true for existing
devices, but likely to be broken in the future.

(cherry picked from commit dadd1a340f735518c198f0a98a925ca2590a9022)

Original-Signed-off-by: Andy Ross <andyross@google.com>
GitOrigin-RevId: dadd1a340f735518c198f0a98a925ca2590a9022
Cr-Build-Id: 8726453333733413585
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8726453333733413585
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Idc45760b5a00c8acf6c40541a4289c6205471895
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6150688
Bot-Commit: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>