drivers/console: Add static buffer feature to winstream_console
The winstream code itself is reasonably generic, but the console
integration was written to the intel_adsp memory window driver only,
as that was the only consumer.

Add an alternative that allows the winstream to be pointed to by a
linker symbol instead.  The platform layer is responsible for putting
it where it needs to go.

Also includes a 20-byte "descriptor" recodr (with a very large magic
number) in the binary that points to the winstream, for the benefit of
external tooling that needs to find it at runtime.

(cherry picked from commit 14418a7f39cecdc2666c7307202ec41d6e3d88bb)

Original-Signed-off-by: Andy Ross <andyross@google.com>
GitOrigin-RevId: 14418a7f39cecdc2666c7307202ec41d6e3d88bb
Cr-Build-Id: 8726453333733413585
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8726453333733413585
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I2bcaebf1b17894d9bc67a1c886d8397d05ea2a67
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6150692
Bot-Commit: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>