drivers: clock_control_nrf2: Add missing cancelation of request
This is a follow-up to commit fe0e2dbc6044ffa74b27bfcba335980bc5a6ae64.

If `nrf_clock_control_request_sync()` ends up with a timeout, before
returning it must cancel the request that was not fulfilled on time.
Otherwise, the request may actually finish successfully a bit later,
but the caller will not be aware that the clock needs to be released
(since the call resulted in an error). And actually even more serious
problem is that because the `req` structure is placed on stack, after
the function returns, the contents of this structure will be probably
overwritten with some other data, so if the request finishes at that
point, an attempt to execute the callback function pointed by this
structure will most likely cause a crash.

(cherry picked from commit c8fff0c77d36628a7b8efd379ca63a1a8483d782)

Original-Signed-off-by: Andrzej Głąbek <andrzej.glabek@nordicsemi.no>
GitOrigin-RevId: c8fff0c77d36628a7b8efd379ca63a1a8483d782
Cr-Build-Id: 8727764308049365969
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8727764308049365969
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I3e56fc72c2e3ae7959cf4236a4203708567a6311
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6115125
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>