soc/mediatek: Don't scan SRAM for winstream in mtk_adsp_load.py
For the same reason that this script can't zero the SRAM region (it
spontaneously reboots my 8186 device) we can't scan the SRAM for the
winstream header.  It's never there anyway, it lives in .rodata.

This is likely a hardware bug, but as it's an older platform and this
is test code it's probably best to just keep it as a workaround.

(cherry picked from commit 619c1ff5778b9bc1cd88a63baaaa20d5886fdcf5)

Original-Signed-off-by: Andy Ross <andyross@google.com>
GitOrigin-RevId: 619c1ff5778b9bc1cd88a63baaaa20d5886fdcf5
Cr-Build-Id: 8726453333733413585
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8726453333733413585
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I3af222d42f3ec80db6f039efa1659c6d57b46ba0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6150694
Commit-Queue: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>