drivers: fuel_gauge: make read-only pointer args const in api
For the lists of properties passed as a pointer to the get_props and
set_props functions, there is no reason to not make the pointer const as
the called functions will not and should not alter the pointed-to data.

In practice, not having the pointers const can cause compilation errors
if trying to pass in a const array of properties.

(cherry picked from commit e4c21da41b4a2ce702daf58bd06882161ca3beb3)

Original-Signed-off-by: Mikkel Jakobsen <mikkel.aunsbjerg@escolifesciences.com>
GitOrigin-RevId: e4c21da41b4a2ce702daf58bd06882161ca3beb3
Cr-Build-Id: 8727631094073597137
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8727631094073597137
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Ic604d2e729552eced1dd562f806b87b1a935456c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6122604
Tested-by: Ting Shen <phoenixshen@chromium.org>
Reviewed-by: Ting Shen <phoenixshen@chromium.org>
Commit-Queue: Ting Shen <phoenixshen@chromium.org>