lib: os: mpsc_pbuf: do not wait when spinlock is held
Check if the spinlock is held before attempting to wait by
taking the semaphore, as that would cause a context switch which
isn't allowed and will trigger an assertion error when
`CONFIG_SPIN_VALIDATE` is enabled.

Logging in spinlock-held context when the log buffer is full can lead
to an infinite assertion error loop, as the logging subsys attempts to
allocate buffer when there's none available, it will try to wait for
one and thus triggers the assertion error, the error message will be
printed through the logging sybsys but there's no buffer available,
so it will try to wait for one and triggers another assertion error..
This loop just goes on and on forever, and nothing gets printed to
the terminal.

Added a test to validate the fix.

(cherry picked from commit 1a578eb72c9dffe74aa3bc47b48400525096218e)

Original-Signed-off-by: Yong Cong Sin <ycsin@meta.com>
Original-Signed-off-by: Yong Cong Sin <yongcong.sin@gmail.com>
Original-Signed-off-by: Maxim Adelman <imax@meta.com>
GitOrigin-RevId: 1a578eb72c9dffe74aa3bc47b48400525096218e
Cr-Build-Id: 8726498632650592545
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8726498632650592545
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I166233de7acb163aaae0a6e65f81792a51aa4b5a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6147343
Commit-Queue: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>