drivers: sensor: Fix TMAG5273/TMAG3001 by adding a TMAG3001 compatible
Fixes issue introduced in #76460

The previous code attempted to detect whether TMAG5273 or TMAG3001 was
connected based on DEVICE ID register. This doesn't work as the bits that
denote the version on one part are undefined on the other part, and cannot
be relied on to be zero.

This commit adds a TMAG3001 compatible which (for now) is just derived from
the TMAG5273 compatible. This allows TMAG3001 to be specified directly in
the DT. The driver code is updated to support both compatibles.

(cherry picked from commit 2bf61e51feb5bb39a7d81f0f9816143eebdf571a)

Original-Signed-off-by: Yiding Jia <yiding.jia@gmail.com>
GitOrigin-RevId: 2bf61e51feb5bb39a7d81f0f9816143eebdf571a
Cr-Build-Id: 8726317439116197601
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8726317439116197601
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Ia407cdaad55c250636ab72d39a5d27e4d1da85b0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6157517
Commit-Queue: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>